creating QueryHandler
